Add Pool usage stats
This commit updates the total, free, and used counts to the pool spec upon
reconcling an IPAddressClaim.

Ints were used to describe these counts, despite the fact that pools can
define ranges much larger than an int. In these cases, math.MaxInt will
be returned. This was done to reduce burden on users of the status.

When Gateway IPs are configured, and they're in the range of the pool,
they are considered in use.

Co-authored-by: Edwin Xie <exie@vmware.com>
Co-authored-by: Aidan Obley <aobley@vmware.com>
Co-authored-by: Christian Ang <angc@vmware.com>